Underlying stream read timeout
If an application is aware of the exact number of bytes a stream will produce, and also reads small bytes in rapid succession towards the end of the stream, an extra stream.Read can occurs before the internal buffers have emptied, causing a timeout pause to can occur. An example of this is streaming to an MSSQL Virtual Device (VDI) interface.