Unmanaged resource leak
If you only call Dispose( ) without Close( ) first, the ZSTD_freeCCtx( ) function is not called, leading to an out of memory error. Also, if the Close( ) function were to throw an error while flushing the ZSTD stream, the memory would also not be released.

By using try/finally on the Close( ), and adding a safety net to Dispose( ) in case Close( ) is not called, the issue seems to be corrected.

Note that if you only call Dispose( ), it will not flush the stream, which does not modify the current behavior. A new function to release resources has been added in order to keep from duplicating code in both Close and DIspose.